requirement.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
blis==0.7.4
catalogue==2.0.6
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.5
click==7.1.2
cymem==2.0.5
dataclasses==0.6
dnspython==2.1.0
elasticsearch==7.14.1
fake-useragent==0.1.11
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
murmurhash==1.0.5
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pathy==0.6.0
preshed==3.0.5
pycares==4.0.0
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==1.2.4
pytz==2021.1
requests==2.26.0
schedule==1.1.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.10
tqdm==4.62.2
twint==2.1.21
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
yarl==1.6.3
